get the spell's damage from getConstantDamage() when calling damage()

+++ src/cpp/auraStats.h
@@ -6,8 +6,8 @@ struct AuraStats
     int hasteRating;
     int shadowPower;
     int firePower;
-    int hastePercent;
+    double hastePercent;
     double manaCostModifier;
 
-    AuraStats(int spellPower, int shadowPower, int firePower, int hasteRating, int hastePercent, double manaCostModifier);
+    AuraStats(int spellPower, int shadowPower, int firePower, int hasteRating, double hastePercent, double manaCostModifier);
 };
